Time: 3 ms (22.93%), Space: 12.3 MB (42.28%) - LeetHub

+++ 0797-rabbits-in-forest/0797-rabbits-in-forest.cpp
@@ -9,7 +9,7 @@ public:
         }
         for(auto x: mp)
         {
-           float q = ceil((float)x.second/(x.first+1));
+           int q = ceil((float)x.second/(x.first+1));
            ans+= (x.first+1)*q;
         }
         return ans;
